requirements.txt
autogen==0.7
openai==1.64.0
python-dotenv==1.1.0
gradio